+++ packages/gossip/manager.go
@@ -45,8 +45,6 @@ type Manager struct {
 	events          Events
 	neighborsEvents map[NeighborsGroup]NeighborsEvents
 
-	wg sync.WaitGroup
-
 	server      *server.TCP
 	serverMutex sync.RWMutex
 
@@ -54,9 +52,9 @@ type Manager struct {
 	neighborsMutex sync.RWMutex
 
 	// messageWorkerPool defines a worker pool where all incoming messages are processed.
-	messageWorkerPool *workerpool.WorkerPool
+	messageWorkerPool *workerpool.NonBlockingQueuedWorkerPool
 
-	messageRequestWorkerPool *workerpool.WorkerPool
+	messageRequestWorkerPool *workerpool.NonBlockingQueuedWorkerPool
 }
 
 // NewManager creates a new Manager.
@@ -76,13 +74,13 @@ func NewManager(local *peer.Local, f LoadMessageFunc, log *logger.Logger) *Manag
 		server:    nil,
 	}
 
-	m.messageWorkerPool = workerpool.New(func(task workerpool.Task) {
+	m.messageWorkerPool = workerpool.NewNonBlockingQueuedWorkerPool(func(task workerpool.Task) {
 		m.processPacketMessage(task.Param(0).([]byte), task.Param(1).(*Neighbor))
 
 		task.Return(nil)
 	}, workerpool.WorkerCount(messageWorkerCount), workerpool.QueueSize(messageWorkerQueueSize))
 
-	m.messageRequestWorkerPool = workerpool.New(func(task workerpool.Task) {
+	m.messageRequestWorkerPool = workerpool.NewNonBlockingQueuedWorkerPool(func(task workerpool.Task) {
 		m.processMessageRequest(task.Param(0).([]byte), task.Param(1).(*Neighbor))
 
 		task.Return(nil)
@@ -97,9 +95,6 @@ func (m *Manager) Start(srv *server.TCP) {
 	defer m.serverMutex.Unlock()
 
 	m.server = srv
-
-	m.messageWorkerPool.Start()
-	m.messageRequestWorkerPool.Start()
 }
 
 // Stop stops the manager and closes all established connections.
@@ -355,6 +350,7 @@ func (m *Manager) processPacketMessage(data []byte, nbr *Neighbor) {
 	packet := new(pb.Message)
 	if err := proto.Unmarshal(data[1:], packet); err != nil {
 		m.log.Debugw("error processing packet", "err", err)
+		return
 	}
 	m.events.MessageReceived.Trigger(&MessageReceivedEvent{Data: packet.GetData(), Peer: nbr.Peer})
 }
@@ -363,16 +359,19 @@ func (m *Manager) processMessageRequest(data []byte, nbr *Neighbor) {
 	packet := new(pb.MessageRequest)
 	if err := proto.Unmarshal(data[1:], packet); err != nil {
 		m.log.Debugw("invalid packet", "err", err)
+		return
 	}
 
 	msgID, _, err := tangle.MessageIDFromBytes(packet.GetId())
 	if err != nil {
 		m.log.Debugw("invalid message id:", "err", err)
+		return
 	}
 
 	msgBytes, err := m.loadMessageFunc(msgID)
 	if err != nil {
 		m.log.Debugw("error loading message", "msg-id", msgID, "err", err)
+		return
 	}
 
 	// send the loaded message directly to the neighbor