+++ packages/ledgerstate/utxo_dag_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/iotaledger/goshimmer/packages/database"
+
 	"github.com/iotaledger/hive.go/crypto/ed25519"
 	"github.com/iotaledger/hive.go/identity"
 	"github.com/iotaledger/hive.go/kvstore/mapdb"
@@ -983,11 +985,12 @@ func TestUTXODAG_CheckTransaction(t *testing.T) {
 
 func setupDependencies(t *testing.T) (*BranchDAG, *UTXODAG) {
 	store := mapdb.NewMapDB()
-	branchDAG := NewBranchDAG(store)
+	cacheTimeProvider := database.NewCacheTimeProvider(0)
+	branchDAG := NewBranchDAG(store, cacheTimeProvider)
 	err := branchDAG.Prune()
 	require.NoError(t, err)
 
-	return branchDAG, NewUTXODAG(store, branchDAG)
+	return branchDAG, NewUTXODAG(store, cacheTimeProvider, branchDAG)
 }
 
 type wallet struct {