+++ packages/txstream/chopper/chopper_test.go
@@ -115,10 +115,11 @@ func TestMaxTangleMessage(t *testing.T) {
 
 //goland:noinspection ALL
 func TestTooLongData(t *testing.T) {
-	tooLong := 1000000
 	short := 10000
 	chunk := 500
-	assert.True(t, tooLong/chunk+1 > 255)
+	tooLong := (chunk * chopper.MaxNChunks)
+
+	assert.True(t, tooLong/chunk >= chopper.MaxNChunks)
 
 	dataLong := make([]byte, tooLong)
 	rand.Read(dataLong)